requirements.txt
pyTelegramBotAPI==4.27.0
telebot==0.0.5
matplotlib==3.10.3
numpy==2.3.1
python-dateutil==2.9.0.post0
requests==2.32.4
pillow==11.3.0
python-dotenv==1.1.1
aiogram==3.4.1